fixes to tests to get them to compile with gcc 4.6, adjustment to makefile to get make clean working in external dirs

+++ test/ccsdt_t3_to_t2.cxx
@@ -112,8 +112,8 @@ int ccsdt_t3_to_t2(int     n,
   int pass = 1;
 
 
-  nrm_AS = sqrt(AS_C["ijkl"]*AS_C["ijkl"]);
-  nrm_NS = sqrt(NS_C["ijkl"]*NS_C["ijkl"]);
+  nrm_AS = sqrt((double)(AS_C["ijkl"]*AS_C["ijkl"]));
+  nrm_NS = sqrt((double)(NS_C["ijkl"]*NS_C["ijkl"]));
 #if DEBUG  >= 1
   if (rank == 0) printf("triangular norm of AS_C = %lf NS_C = %lf\n", nrm_AS, nrm_NS);
 #endif
